Fix Avatar crash on complicated unicode names

+++ ios/FluentUI/Avatar/Avatar.swift
@@ -433,7 +433,7 @@ public struct Avatar: View, TokenizedControlView {
             var hash: Int32 = 0
             for len in (0..<text.length).reversed() {
                 let ch = text.character(at: len)
-                let shift = len % 8
+                let shift = Int32(len % 8)
                 hash ^= Int32((ch << shift) + (ch >> (8 - shift)))
               }
